+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### 0.1.2
+- Properly added ChangePlayerSize action with different modes, and framework to easily add more "change x" actions.
 - Separated MapScale from LevelScale, this is useful for re-using maps, editing big levels, and more in the future.
 
 ### 0.1.1